moved gcs workspace index to backend
